# Switched to in-place update of the diagonal Hessian