--- a/gtsam/linear/HessianFactor.cpp
+++ b/gtsam/linear/HessianFactor.cpp
@@ -315,6 +315,7 @@ void HessianFactor::hessianDiagonalAdd(VectorValues &d) const {
   for (DenseIndex j = 0; j < (DenseIndex)size(); ++j) {
     auto result = d.emplace(keys_[j], info_.diagonal(j));
     if(!result.second) {
+      // if emplace fails, it returns an iterator to the existing element, which we add to:
       result.first->second += info_.diagonal(j);
     }
   }