--- a/gtsam/linear/linearAlgorithms-inst.h
+++ b/gtsam/linear/linearAlgorithms-inst.h
@@ -100,9 +100,9 @@ namespace gtsam
             for(GaussianConditional::const_iterator frontal = c.beginFrontals(); frontal != c.endFrontals(); ++frontal) {
               auto result = collectedResult.emplace(*frontal, solution.segment(vectorPosition, c.getDim(frontal)));
               if(!result.second)
-                  throw std::invalid_argument(
-                      "Requested to emplace variable '" + DefaultKeyFormatter(*frontal)
-                      + "' already in this VectorValues.");
+                  throw std::runtime_error(
+                      "Internal error while optimizing clique. Trying to insert key '" + DefaultKeyFormatter(*frontal)
+                      + "' that exists.");
 
               VectorValues::const_iterator r = result.first;
               myData.cliqueResults.emplace(r->first, r);