--- a/gtsam/slam/RegularImplicitSchurFactor.h
+++ b/gtsam/slam/RegularImplicitSchurFactor.h
@@ -147,7 +147,7 @@ public:
     VectorValues d;
 
     hessianDiagonalAdd(d);
-    
+
     return d;
   }