Benchmark with `SEQUENTIAL_CHOLESKY`

Warmed-up timing
================

Before:
```
Initial error: 4.18566e+06, values: 22122
iter      cost      cost_change    lambda  success iter_time
   0  1.030608e+05    4.08e+06    1.00e-04     1    7.43e+00
   1  6.149065e+04    4.16e+04    3.33e-05     1    7.27e+00
   2  1.956166e+04    4.19e+04    3.33e-05     1    7.28e+00
   3  1.814774e+04    1.41e+03    1.11e-05     1    7.31e+00
   4  1.803417e+04    1.14e+02    3.98e-06     1    7.34e+00
   5  1.803390e+04    2.62e-01    1.33e-06     1    7.35e+00
   6  1.803390e+04    8.14e-05    4.42e-07     1    7.36e+00
-Total: 0 CPU (0 times, 0 wall, 74.61 children, min: 0 max: 0)
|   -optimize: 74.61 CPU (1 times, 74.7746 wall, 74.61 children, min: 74.61 max: 74.61)
LD_LIBRARY_PATH=/opt/intel/lib ../gtsam_build/timing/timeSFMBAL  74.23s user 0.88s system 99% cpu 1:15.27 total
```

After:
```
Initial error: 4.18566e+06, values: 22122
iter      cost      cost_change    lambda  success iter_time
   0  1.030608e+05    4.08e+06    1.00e-04     1    7.47e+00
   1  6.149065e+04    4.16e+04    3.33e-05     1    7.30e+00
   2  1.956166e+04    4.19e+04    3.33e-05     1    7.29e+00
   3  1.814774e+04    1.41e+03    1.11e-05     1    7.29e+00
   4  1.803417e+04    1.14e+02    3.98e-06     1    7.32e+00
   5  1.803390e+04    2.62e-01    1.33e-06     1    7.35e+00
   6  1.803390e+04    8.14e-05    4.42e-07     1    7.37e+00
-Total: 0 CPU (0 times, 0 wall, 71.2 children, min: 0 max: 0)
|   -optimize: 71.2 CPU (1 times, 71.3531 wall, 71.2 children, min: 71.2 max: 71.2)
LD_LIBRARY_PATH=/opt/intel/lib ../gtsam_build/timing/timeSFMBAL  70.95s user 0.74s system 99% cpu 1:11.84 total
```

Cold timing
===============

Before:
```
Initial error: 4.18566e+06, values: 22122
iter      cost      cost_change    lambda  success iter_time
   0  1.030608e+05    4.08e+06    1.00e-04     1    1.03e+01
   1  6.149065e+04    4.16e+04    3.33e-05     1    8.06e+00
   2  1.956166e+04    4.19e+04    3.33e-05     1    8.07e+00
   3  1.814774e+04    1.41e+03    1.11e-05     1    7.77e+00
   4  1.803417e+04    1.14e+02    3.98e-06     1    8.16e+00
   5  1.803390e+04    2.62e-01    1.33e-06     1    7.77e+00
   6  1.803390e+04    8.14e-05    4.42e-07     1    7.80e+00
-Total: 0 CPU (0 times, 0 wall, 82.97 children, min: 0 max: 0)
|   -optimize: 82.97 CPU (1 times, 85.3366 wall, 82.97 children, min: 82.97 max: 82.97)
LD_LIBRARY_PATH=/opt/intel/lib ../gtsam_build/timing/timeSFMBAL  81.95s user 1.64s system 97% cpu 1:26.04 total
```

After:

```
Initial error: 4.18566e+06, values: 22122
iter      cost      cost_change    lambda  success iter_time
   0  1.030608e+05    4.08e+06    1.00e-04     1    8.29e+00
   1  6.149065e+04    4.16e+04    3.33e-05     1    8.29e+00
   2  1.956166e+04    4.19e+04    3.33e-05     1    7.48e+00
   3  1.814774e+04    1.41e+03    1.11e-05     1    7.69e+00
   4  1.803417e+04    1.14e+02    3.98e-06     1    7.83e+00
   5  1.803390e+04    2.62e-01    1.33e-06     1    7.48e+00
   6  1.803390e+04    8.14e-05    4.42e-07     1    7.58e+00
-Total: 0 CPU (0 times, 0 wall, 75.98 children, min: 0 max: 0)
|   -optimize: 75.98 CPU (1 times, 76.699 wall, 75.98 children, min: 75.98 max: 75.98)
LD_LIBRARY_PATH=/opt/intel/lib ../gtsam_build/timing/timeSFMBAL  75.47s user 1.07s system 98% cpu 1:17.34 total
```

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/borglab/gtsam/337)
<!-- Reviewable:end -->