@ProfFan 2020-06-02
It’s 3%-4% of the SEQUENTIAL_CHOLESKY time, should be more with EIGEN_CHOLESKY