@ProfFan 2020-06-03
There are some strange errors:
```
94/181 Test  #94: testJacobianFactor .................***Failed    0.09 sec
Not equal:
expected:
: 3 elements
  5: 1 1 1
  10: 4 4 4
  15: 9 9 9
actual:
: 3 elements
  5: 0.999998 0.999998 0.999998
  10: 4 4 4
  15: 9 9 9
 98/181 Test  #98: testRegularJacobianFactor ..........***Failed    0.08 sec
Not equal:
expected:
: 3 elements
  0: 4 4 4
  1: 16 16 16
  2: 36 36 36
actual:
: 3 elements
  0: 4 4 4
  1: 16 16 16
  2: 36 36 36
```