@dellaert 2020-06-03
Unit test emplace

On Tue, Jun 2, 2020 at 21:44 Fan Jiang <notifications@github.com> wrote:

> There are some strange errors:
>
> 94/181 Test  #94: testJacobianFactor .................***Failed    0.09 sec
> Not equal:
> expected:
> : 3 elements
>   5: 1 1 1
>   10: 4 4 4
>   15: 9 9 9
> actual:
> : 3 elements
>   5: 0.999998 0.999998 0.999998
>   10: 4 4 4
>   15: 9 9 9
>  98/181 Test  #98: testRegularJacobianFactor ..........***Failed    0.08 sec
> Not equal:
> expected:
> : 3 elements
>   0: 4 4 4
>   1: 16 16 16
>   2: 36 36 36
> actual:
> : 3 elements
>   0: 4 4 4
>   1: 16 16 16
>   2: 36 36 36
>
> —
> You are receiving this because you commented.
> Reply to this email directly, view it on GitHub
> <https://github.com/borglab/gtsam/pull/337#issuecomment-637903878>, or
> unsubscribe
> <https://github.com/notifications/unsubscribe-auth/ACQHGSMUGEFZ43LT6MG22GTRUWTGXANCNFSM4NQ4XUOA>
> .
>
-- 
Best !
Frank Dellaert
http://frank.dellaert.com