@ProfFan 2020-06-03
@dellaert Fixed. Now all unit tests should pass.