@dellaert 2020-06-03
New timing?