@dellaert 2020-06-03
> ```
> -Total: 0 CPU (0 times, 0 wall, 73.65 children, min: 0 max: 0)
> |   -optimize: 73.65 CPU (1 times, 73.9181 wall, 73.65 children, min: 73.65 max: 73.65)
> LD_LIBRARY_PATH=/opt/intel/lib ../gtsam_build/timing/timeSFMBAL  73.23s user 0.94s system 99% cpu 1:14.43 total
> ```
> 
> @dellaert 1:14 with `SEQUENTIAL_CHOLESKY` :)

Wait, I don't understand. It's 73s now, compared to 82s when you started the PR (yes!!!! > 10% saving !?), but with what solver? Note Eigen is not in develop yet, so the description of this PR should state the savings for using an existing solver, e.g. SEQUENTIAL_CHOLESKY. Could you update the PR description with a before and after for SEQUENTIAL_CHOLESKY?