@ProfFan 2020-06-03
@dellaert I updated the benchmark with proper warm-up (for filling CPU cache). Can see about 1s of improvement in runtime.