@dellaert 2020-06-03
> @dellaert I updated the benchmark with proper warm-up (for filling CPU cache). Can see about 1s of improvement in runtime.

Hmm. That's disappointing. But, on the other hand, why is "warm-started?" the right benchmark? People typically run optimizations once.