@ProfFan 2020-06-03
@dellaert Because that will eliminate the disturbance caused by other processes filling the CPU cache, etc. https://engineering.appfolio.com/appfolio-engineering/2017/5/2/what-about-warmup