@ProfFan 2020-06-03
I did another optimization to further reduce heap allocation. Now the benchmark:
```
Initial error: 4.18566e+06, values: 22122
iter      cost      cost_change    lambda  success iter_time
   0  1.030608e+05    4.08e+06    1.00e-04     1    7.47e+00
   1  6.149065e+04    4.16e+04    3.33e-05     1    7.30e+00
   2  1.956166e+04    4.19e+04    3.33e-05     1    7.29e+00
   3  1.814774e+04    1.41e+03    1.11e-05     1    7.29e+00
   4  1.803417e+04    1.14e+02    3.98e-06     1    7.32e+00
   5  1.803390e+04    2.62e-01    1.33e-06     1    7.35e+00
   6  1.803390e+04    8.14e-05    4.42e-07     1    7.37e+00
-Total: 0 CPU (0 times, 0 wall, 71.2 children, min: 0 max: 0)
|   -optimize: 71.2 CPU (1 times, 71.3531 wall, 71.2 children, min: 71.2 max: 71.2)
LD_LIBRARY_PATH=/opt/intel/lib ../gtsam_build/timing/timeSFMBAL  70.95s user 0.74s system 99% cpu 1:11.84 total
```