@dellaert 2020-06-03
@ProfFan awesome! But (a) could you update the description with *cold* timing? (b) the build seems to be failing :-/