@ProfFan 2020-06-03
@dellaert Already done, see top