@ProfFan 2020-06-03
BTW, in profile the amount of total execution time consumed by `hessianDiagonal()` reduced from 10% to 5%.