@ProfFan 2020-06-03
@dellaert  It's possible but not trivial - `VectorValues` is forward declared in `GaussianFactor`, so we cannot use it for real. Also `GaussianFactor` is header-only so we cannot put implementations.