@ProfFan 2020-06-04
@dellaert Done.