@dellaert 2020-06-04
Awesome ! Let's do some SwiftFusion now, pick up Eigen back up after we meet w those folks :-)